+++ test/svlog/regr/iss206_unsup_sysfunc.sv
@@ -1,0 +1,7 @@
+// RUN: moore %s -e foo
+// FAIL
+
+module foo;
+    initial $bar();
+    // CHECK-ERR: error: unknown system task `$bar`
+endmodule